Rewrite create base config,write test whose check connection to db
Post title now use css text-weight property bold

+++ Flask_blog/config_loader.py
@@ -7,10 +7,15 @@ def create_base_config(path):
     """Create configuration file
     :arg path path to config file
     """
-    config = f"secret_key={os.urandom(20).hex()}\n" \
-             f"env=production"
-    with open(path, "w") as file:
-        file.write(config)
+    if os.path.isfile(path):
+        print("Config file already exist")
+        exit(0)
+    else:
+        config = f"secret_key={os.urandom(20).hex()}\n" \
+                 f"env=production\n" \
+                 f"MONGO_URI='paste adress of your database here'"
+        with open(path, "w") as file:
+            file.write(config)
 
 
 def load_environment_variables(path_to_file=".env"):
@@ -19,3 +24,6 @@ def load_environment_variables(path_to_file=".env"):
         load_dotenv(dotenv_path=path_to_file)
     except FileExistsError:
         raise FileExistsError("Environment file not found")
+
+if __name__ == "__main__":
+    create_base_config(".env")
